# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Admin {
  id: number;
  email: string;
  name: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AuthSession {
  admin: Admin;
}

export type StudentLevel = typeof StudentLevel[keyof typeof StudentLevel];


export const StudentLevel = {
  A1: 'A1',
  A2: 'A2',
  B1: 'B1',
  B2: 'B2',
} as const;

export type StudentStatus = typeof StudentStatus[keyof typeof StudentStatus];


export const StudentStatus = {
  training: 'training',
  ready: 'ready',
  placed: 'placed',
} as const;

export type LevelHistoryLevel = typeof LevelHistoryLevel[keyof typeof LevelHistoryLevel];


export const LevelHistoryLevel = {
  A1: 'A1',
  A2: 'A2',
  B1: 'B1',
  B2: 'B2',
} as const;

export interface LevelHistory {
  level: LevelHistoryLevel;
  completedAt: string;
  /** @nullable */
  score?: number | null;
}

export type PlacementProgram = typeof PlacementProgram[keyof typeof PlacementProgram];


export const PlacementProgram = {
  Ausbildung: 'Ausbildung',
  Arbeit: 'Arbeit',
  Praktikum: 'Praktikum',
} as const;

export interface Placement {
  country: string;
  program: PlacementProgram;
  company?: string;
  placedAt: string;
}

export interface Student {
  id: number;
  name: string;
  initials: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  photoUrl: string | null;
  level: StudentLevel;
  status: StudentStatus;
  cohort: string;
  joinedAt: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  speakingVideoUrl?: string | null;
  /** @nullable */
  certificateName?: string | null;
  /** @nullable */
  certificateUrl?: string | null;
  levelHistory: LevelHistory[];
  placement?: Placement | null;
}

export type StudentInputLevel = typeof StudentInputLevel[keyof typeof StudentInputLevel];


export const StudentInputLevel = {
  A1: 'A1',
  A2: 'A2',
  B1: 'B1',
  B2: 'B2',
} as const;

export type StudentInputStatus = typeof StudentInputStatus[keyof typeof StudentInputStatus];


export const StudentInputStatus = {
  training: 'training',
  ready: 'ready',
  placed: 'placed',
} as const;

export interface StudentInput {
  /** @minLength 2 */
  name: string;
  email: string;
  phone?: string;
  photoUrl?: string;
  level: StudentInputLevel;
  status: StudentInputStatus;
  cohort: string;
  joinedAt: string;
  bio?: string;
  speakingVideoUrl?: string;
  certificateName?: string;
  certificateUrl?: string;
  placement?: Placement;
}

export type StudentUpdateLevel = typeof StudentUpdateLevel[keyof typeof StudentUpdateLevel];


export const StudentUpdateLevel = {
  A1: 'A1',
  A2: 'A2',
  B1: 'B1',
  B2: 'B2',
} as const;

export type StudentUpdateStatus = typeof StudentUpdateStatus[keyof typeof StudentUpdateStatus];


export const StudentUpdateStatus = {
  training: 'training',
  ready: 'ready',
  placed: 'placed',
} as const;

export interface StudentUpdate {
  /** @minLength 2 */
  name?: string;
  email?: string;
  phone?: string;
  photoUrl?: string;
  level?: StudentUpdateLevel;
  status?: StudentUpdateStatus;
  cohort?: string;
  joinedAt?: string;
  bio?: string;
  speakingVideoUrl?: string;
  certificateName?: string;
  certificateUrl?: string;
  placement?: Placement;
}

export type LevelCountLevel = typeof LevelCountLevel[keyof typeof LevelCountLevel];


export const LevelCountLevel = {
  A1: 'A1',
  A2: 'A2',
  B1: 'B1',
  B2: 'B2',
} as const;

export interface LevelCount {
  level: LevelCountLevel;
  count: number;
}

export interface PeriodStat {
  period: string;
  total: number;
  placed: number;
}

export interface DashboardSummary {
  totalStudents: number;
  readyToPlace: number;
  placedStudents: number;
  placementRate: number;
  byLevel: LevelCount[];
  byPeriod: PeriodStat[];
  recentStudents: Student[];
}

export type ListStudentsParams = {
level?: ListStudentsLevel;
status?: ListStudentsStatus;
cohort?: string;
search?: string;
};

export type ListStudentsLevel = typeof ListStudentsLevel[keyof typeof ListStudentsLevel];


export const ListStudentsLevel = {
  A1: 'A1',
  A2: 'A2',
  B1: 'B1',
  B2: 'B2',
} as const;

export type ListStudentsStatus = typeof ListStudentsStatus[keyof typeof ListStudentsStatus];


export const ListStudentsStatus = {
  training: 'training',
  ready: 'ready',
  placed: 'placed',
} as const;