# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LevelCount } from './levelCount';
import type { PeriodStat } from './periodStat';
import type { Student } from './student';

export interface DashboardSummary {
  totalStudents: number;
  readyToPlace: number;
  placedStudents: number;
  placementRate: number;
  byLevel: LevelCount[];
  byPeriod: PeriodStat[];
  recentStudents: Student[];
}