# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LevelHistory } from './levelHistory';
import type { Placement } from './placement';
import type { StudentLevel } from './studentLevel';
import type { StudentStatus } from './studentStatus';

export interface Student {
  id: number;
  name: string;
  initials: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  photoUrl: string | null;
  level: StudentLevel;
  status: StudentStatus;
  cohort: string;
  joinedAt: Date;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  speakingVideoUrl?: string | null;
  /** @nullable */
  certificateName?: string | null;
  /** @nullable */
  certificateUrl?: string | null;
  levelHistory: LevelHistory[];
  placement?: Placement | null;
}