# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Placement } from './placement';
import type { StudentInputLevel } from './studentInputLevel';
import type { StudentInputStatus } from './studentInputStatus';

export interface StudentInput {
  /** @minLength 2 */
  name: string;
  email: string;
  phone?: string;
  photoUrl?: string;
  level: StudentInputLevel;
  status: StudentInputStatus;
  cohort: string;
  joinedAt: Date;
  bio?: string;
  speakingVideoUrl?: string;
  certificateName?: string;
  certificateUrl?: string;
  placement?: Placement;
}