# lib/api-zod/src/generated/types/studentUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Placement } from './placement';
import type { StudentUpdateLevel } from './studentUpdateLevel';
import type { StudentUpdateStatus } from './studentUpdateStatus';

export interface StudentUpdate {
  /** @minLength 2 */
  name?: string;
  email?: string;
  phone?: string;
  photoUrl?: string;
  level?: StudentUpdateLevel;
  status?: StudentUpdateStatus;
  cohort?: string;
  joinedAt?: Date;
  bio?: string;
  speakingVideoUrl?: string;
  certificateName?: string;
  certificateUrl?: string;
  placement?: Placement;
}